organize sections into chapters
