+++ game/scripts/vscripts/Sections/Chapter1/SectionCameraUnlock.ts
@@ -1,7 +1,7 @@
-import * as tg from "../TutorialGraph/index"
-import * as tut from "../Tutorial/Core"
-import { getOrError, getPlayerHero, setUnitPacifist } from "../util"
-import { TutorialContext } from "../TutorialGraph/index"
+import * as tg from "../../TutorialGraph/index"
+import * as tut from "../../Tutorial/Core"
+import { getOrError, getPlayerHero, setUnitPacifist } from "../../util"
+import { TutorialContext } from "../../TutorialGraph/index"
 
 let graph: tg.TutorialStep | undefined = undefined
