+++ game/scripts/vscripts/Sections/Chapter1/SectionCasting.ts
@@ -1,6 +1,6 @@
-import * as tg from "../TutorialGraph/index";
-import * as tut from "../Tutorial/Core";
-import { getPlayerHero } from "../util";
+import * as tg from "../../TutorialGraph/index";
+import * as tut from "../../Tutorial/Core";
+import { getPlayerHero } from "../../util";
 
 let graph: tg.TutorialStep | undefined = undefined;
 let graphContext: tg.TutorialContext | undefined = undefined;
