+++ game/scripts/vscripts/Sections/Chapter1/SectionOpening.ts
@@ -1,6 +1,6 @@
-import * as tg from "../TutorialGraph/index"
-import * as tut from "../Tutorial/Core"
-import { getPlayerHero } from "../util"
+import * as tg from "../../TutorialGraph/index"
+import * as tut from "../../Tutorial/Core"
+import { getPlayerHero } from "../../util"
 
 let graph: tg.TutorialStep | undefined = undefined
