+++ game/scripts/vscripts/Sections/Chapter3/index.ts
@@ -1,0 +1,2 @@
+
+export * from "./Chapter3"
