+++ game/scripts/vscripts/Sections/index.ts
@@ -1,5 +1,2 @@
-export * from "./SectionOpening"
-export * from "./SectionCameraUnlock"
-export * from "./SectionCasting"
-export * from "./SectionLeveling"
-export * from "./Chapter3/Chapter3"
+export * as chapter1 from "./Chapter1/index"
+export * as chapter3 from "./Chapter3/index"
